Unregister flat product entity after tests — refs #49
It is required as running tests in isolated processes does not seem to work
and if the test uses data retrieval on different stores, the product entity associated
with the first store will be kept in the registry, leading to wrong results with
flat catalog enabled

Ex: in a test class, a first test retrieves products from store id 1, the second
test retrieves products from store id 2. Because of the object remaining in the
registry between those 2 calls, the second call will still query against table
catalog_product_flat_1, the object in the regitry having it store id set to 1
from the first test

It's not required to unregister in the Callback class, as only tests may successively
call on different store. The real ShopyMind client will have 1 store per HTTP call